[case](regression-test) load json with enable_simdjson_reader=false
