+++ regression-test/data/load_p0/stream_load/test_json_load.out
@@ -205,9 +205,16 @@ John	30	New York	{"email":"john@example.com","phone":"+1-123-456-7890"}
 -- !select22 --
 11324	1321313082437	1678834024274	20230315	{"base_mac_value_null":24,"base_1_value_respiratoryrate":11,"base_3_value_heartrate":51,"base_3_status_onoroutofbed":3,"base_null_count_circulation":84,"base_1_status_onoroutofbed":3,"base_1_value_heartrate":51,"base_3_value_respiratoryrate":11,"base_3_value_bodyactivityenergy":43652,"base_2_value_respiratoryrate":11,"base_2_value_bodyactivityenergy":28831,"base_2_status_onoroutofbed":3,"base_1_value_bodyactivityenergy":56758,"base_2_value_heartrate":51,"tsltype":"properties","datatimestamp":1678834024274,"command":"0105","macaddress":"405EE1805029"}
 
--- !select --
-200	changsha	3456789
-
--- !select --
-200	changsha	3456789
+-- !select23 --
+10	1454547
+20	1244264
+30	528369
+40	594201
+50	594201
+60	2345672
+70	2345673
+80	2345674
+90	2345675
+100	2345676
+200	755
